Make xml serializers depend on AbstractExtensionXmlSerializer to avoid duplication and put tests in an independent class
